task 7 edited

+++ src/lesson7/main.java
@@ -8,18 +8,18 @@ public class main {
         System.out.println("Enter symbol: ");
         String symbol = sc.nextLine();
         System.out.println("Enter your number: ");
-        double number = sc.nextDouble();
+        int number = sc.nextInt();
 
         System.out.println(getCube(number));
         printStar(number);
         printSymbolNumberTimes(number, symbol);
     }
 
-    public static double getCube(double a) {
-        return Math.pow(a, 3);
+    public static int getCube(int a) {
+        return (int) Math.pow(a, 3);
     }
 
-    public static void printStar(double a) {
+    public static void printStar(int a) {
         while (a > 0) {
             System.out.print("*");
             a--;
@@ -27,8 +27,8 @@ public class main {
         System.out.println();
     }
 
-    public static void printSymbolNumberTimes(double number, String symbol) {
-        for (int i = 0; i < (int) number; i++) {
+    public static void printSymbolNumberTimes(int number, String symbol) {
+        for (int i = 0; i < number; i++) {
             System.out.print(symbol);
         }
     }
